Relocate LibraryTab method trackOpenNewDatabase() to OpenDatabaseAction
